% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms2viirs.R
\name{vms2viirs}
\alias{vms2viirs}
\title{vms 2 viirs}
\usage{
vms2viirs(viirsfile, vmsfile, ROI)
}
\arguments{
\item{viirsfile}{is a geospatial dataframe with viirs data obtained from https://eogdata.mines.edu/vbd/}

\item{vmsfile}{is a geospatial dataframe with VMS data over a certain period of time obtained via KKP (the Fisheries Ministrie of Indonesia)}

\item{is}{a geospatial dataframe with a Region of Interest which determines the scope of your research area.}
}
\value{
a dataframe with positions of vessels that were included in the VMS dataset at the time of overpass of the VIIRS satellite.
}
\description{
vms 2 viirs
}
