% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms2viirs2shp.R
\name{vms2viirs2shp}
\alias{vms2viirs2shp}
\title{vms 2 viirs 2 shp}
\usage{
vms2viirs2shp(vms2viirs, ROIfile)
}
\arguments{
\item{vms2viirs}{is the output of the earlier 'vms2viirs' function}

\item{ROIfile}{is a geospatial dataframe with a Region of Interest which determines the scope of your research area.}
}
\value{
the output of vms2viirs in s4 format (can be exported to shapefile now)
}
\description{
vms 2 viirs 2 shp
}
