% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vms2viirs4path.R
\name{vms2viirs4path}
\alias{vms2viirs4path}
\title{vms 2 viirs 4 path}
\usage{
vms2viirs4path(vmsanalysisfile, vmsfile)
}
\arguments{
\item{vmsanalysisfile}{is the output of the 'vms2viirsanalysis' function.}

\item{vmsfile}{is a geospatial dataframe with VMS data over a certain period of time obtained via KKP (the Fisheries Ministrie of Indonesia)}
}
\value{
a daily path file for all detected vessels of the vms2viirsanalysis function.
}
\description{
vms 2 viirs 4 path
}
